3 - get back to Int type for user's id

+++ src/main/scala/tech/cryptonomic/nautilus/cloud/domain/user/User.scala
@@ -18,5 +18,5 @@ case class User(
 }
 
 object User {
-  type UserId = Long
+  type UserId = Int
 }
